Depend on WPEBackend.
The first issue to be faced is that you have to install WPEWebKit into
system directories. It doesn't seem like you can depend on a locally
built WPEWebKit, which would be nice for development (to avoid having to
install the project into /usr after each change). This is because the
pkg-config template in WebKit has /usr/local as the library prefix.

The second issue I had after install WPEWebKit was this,

[ 16%] Building C object CMakeFiles/dinghycore.dir/dy-launcher.c.o
In file included from /usr/local/include/wpe-0.1/WPE/wpe/WebKitWebView.h:52:0,
                 from /usr/local/include/wpe-0.1/WPE/wpe/webkit.h:74,
                 from /home/charlie/git/dinghy/dy-webkit-utils.h:21,
                 from /home/charlie/git/dinghy/dy-launcher.h:16,
                 from /home/charlie/git/dinghy/dy-launcher.c:8:
/usr/local/include/wpe-0.1/WPE/wpe/WebKitWebViewBackend.h:29:10: fatal error: wpe/view-backend.h: No such file or directory
 #include <wpe/view-backend.h>
          ^~~~~~~~~~~~~~~~~~~~

This is caused because dingy does not depend on WPEBackend. WPEBackend
is build as part of the WPE dependencies jhbuild, and there doesn't seem
to be support for installing those into system directories. Unlike the
WebKit pkg-config, these ones have a local pkg-config path (like
/home/charlie/WebKit/WebKit/WebKitBuild/DependenciesWPE/Root).

This patch mainly tries to fix this missing dependency, the other bits
of the diff are just fixing a typo.

I can now run dingy, but I have to do things like

LD_LIBRARY_PATH=/usr/local/lib:/home/charlie/WebKit/WebKit/WebKitBuild/DependenciesWPE/Root/lib
./dinghy

To pick up both the WebKit libs and the WPEBackend libs. The
cmake_install.cmake files in the jhbuild deps only installs locally, not
system-wide. So without some rpath trickery, I'm not sure what is a
better way of working with dingy.